project file that works for both of the input files given for the assignment
Added detailed comments for each function showing the instruction she gave use for each function, the notes I took form the lab and the thought process used for the more difficult functions. I also re-ordered the case statements in the instruction_decode function so they would be in ascending order, however, the only change I made to the control signals was set the Don't Care signals to 2.